Fixing lint errors

+++ tests/spec/unit_tests.js
@@ -19,60 +19,56 @@
 /* eslint-env es6, node */
 /* eslint-disable no-console */
 
-'use strict';
-const isNodeProcess = (typeof process === 'object' && process + '' === '[object process]');
+const isNodeProcess = typeof process === "object" && process + "" === "[object process]";
 
-const fs = require('fs');
-const path = require('path');
-const Realm = require('realm');
+const fs = require("fs");
+const path = require("path");
+const Realm = require("realm");
 
 // Setup a Realm logger
-const RealmLogging = require('./helpers/realm-logging');
+const RealmLogging = require("./helpers/realm-logging");
 RealmLogging.patch(Realm);
 
-const RealmTests = require('../js');
+const RealmTests = require("../js");
 
 jasmine.DEFAULT_TIMEOUT_INTERVAL = 30000;
-let isDebuggerAttached = typeof v8debug === 'object';
+let isDebuggerAttached = typeof v8debug === "object";
 if (!isDebuggerAttached && isNodeProcess) {
-    isDebuggerAttached = /--debug|--inspect/.test(process.execArgv.join(' '));
+  isDebuggerAttached = /--debug|--inspect/.test(process.execArgv.join(" "));
 }
 
 if (isDebuggerAttached) {
-    jasmine.DEFAULT_TIMEOUT_INTERVAL = 3000000;
+  jasmine.DEFAULT_TIMEOUT_INTERVAL = 3000000;
 }
 
 // Create this method with appropriate implementation for Node testing.
 Realm.copyBundledRealmFiles = function () {
-    let sourceDir = path.join(__dirname, '../data');
-    let destinationDir = path.dirname(Realm.defaultPath);
+  let sourceDir = path.join(__dirname, "../data");
+  let destinationDir = path.dirname(Realm.defaultPath);
 
-    for (let filename of fs.readdirSync(sourceDir)) {
-        let src = path.join(sourceDir, filename);
-        let dest = path.join(destinationDir, filename);
+  for (let filename of fs.readdirSync(sourceDir)) {
+    let src = path.join(sourceDir, filename);
+    let dest = path.join(destinationDir, filename);
 
-        // If the destination file already exists, then don't overwrite it.
-        try {
-            fs.accessSync(dest);
-            continue;
-        } catch (e) { }
-
-        fs.writeFileSync(dest, fs.readFileSync(src));
+    // If the destination file already exists, then don't overwrite it.
+    const exists = fs.existsSync(dest);
+    if (!exists) {
+      fs.writeFileSync(dest, fs.readFileSync(src));
     }
+  }
 };
 
 const tests = RealmTests.getTestNames();
 for (const suiteName in tests) {
-    describe(suiteName, () => {
+  describe(suiteName, () => {
+    beforeAll((done) => RealmTests.prepare(done));
 
-        beforeAll(done => RealmTests.prepare(done));
+    beforeEach(() => RealmTests.runTest(suiteName, "beforeEach"));
 
-        beforeEach(() => RealmTests.runTest(suiteName, 'beforeEach'));
+    for (const testName of tests[suiteName]) {
+      it(testName, () => RealmTests.runTest(suiteName, testName));
+    }
 
-        for (const testName of tests[suiteName]) {
-            it(testName, () => RealmTests.runTest(suiteName, testName));
-        }
-
-        afterEach(() => RealmTests.runTest(suiteName, 'afterEach'));
-    });
+    afterEach(() => RealmTests.runTest(suiteName, "afterEach"));
+  });
 }
